* commit '0ce7a6abadd227d199d0ce66c5f8220561c96974':
  Fix all phpstan level 3 and level 2 errors

+++ src/Storefront/Page/Listing/AggregationView/SliderAggregation.php
@@ -72,7 +72,7 @@ class SliderAggregation extends Struct implements AggregationViewInterface
         string $minFieldName,
         string $maxFieldName,
         ?string $suffix = null,
-        float $digits = 2
+        int $digits = 2
     ) {
         $this->name = $name;
         $this->active = $active;
